K8SPSMDB-1197: fix healthcheck log rotation
https://perconadev.atlassian.net/browse/K8SPSMDB-1197
